pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "arcam-fmj"
version = "2.2.0"
description = "A python library for speaking to Arcam receivers"
readme = "README.rst"
authors = [
    {name = "Joakim Plate"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Plugins",
    "Framework :: AsyncIO",
]
requires-python = ">=3.11"
dependencies = [
    "attrs>=22.1",
]
license = "MIT"

[project.urls]
Homepage = "https://github.com/elupus/arcam_fmj"

[project.optional-dependencies]
discovery = [
    "aiohttp>=3.8",
    "defusedxml>=0.7.1",
]
cli = [
    "rich>=13.0",
]
tests = [
    "arcam-fmj[discovery]",
    "pytest>=8.0",
    "pytest-aiohttp>=1.0.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=3.0.0",
    "pytest-mock>=3.10",
]

[project.scripts]
arcam-fmj = "arcam.fmj.console:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]

[tool.ruff.lint.per-file-ignores]
"tests/test_utils.py" = ["E501"]  # XML test fixtures contain long lines

[tool.ruff.lint.isort]
known-first-party = ["arcam"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["defusedxml.*", "aiohttp.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["arcam"]

[tool.coverage.report]
fail_under = 85
